cursor: Move cursor specific function definitions into cursor.h
